Theo code commit

+++ PROJET_SNAKE_Archive_Eleve/myLib.c
@@ -251,7 +251,7 @@ void setBackgroundColor(int c) {
 	CONSOLE_SCREEN_BUFFER_INFO csbi;
 
 	GetConsoleScreenBufferInfo(hConsole, &csbi);
-	const DWORD dwConSize = (unsigned long) (csbi.dwSize.X * csbi.dwSize.Y);
+	const DWORD dwConSize = csbi.dwSize.X * csbi.dwSize.Y;
 	FillConsoleOutputCharacter(hConsole, (TCHAR)' ', dwConSize, coordScreen, &cCharsWritten);
 
 	GetConsoleScreenBufferInfo(hConsole, &csbi);
@@ -270,7 +270,7 @@ void setBackgroundColor(int c) {
 int random(){
     static int k=0;
     if (k==0){
-            srand((unsigned int)time(NULL));
+            srand(time(NULL));
             k=1;}
 
     return rand();
@@ -288,12 +288,12 @@ void printSnakeHead(){
     printf("%c",FORME_TETE);
 }
 
-/* Affiche un ï¿½lï¿½ment du corps du serpent*/
+/* Affiche un élément du corps du serpent*/
 void printSnakeBody(){
     printf("%c",FORME_CORPS);
 }
 
-/* Affiche un ï¿½lï¿½ment du stade */
+/* Affiche un élément du stade */
 void printStadeElement(){
     printf("%c",FORME_STADE);
 }
